move all useful classes to kernel package
