+++ src/main/scala/ammonite/kernel/Name.scala
@@ -1,4 +1,4 @@
-package ammonite.util
+package ammonite.kernel
 
 import scala.reflect.NameTransformer
