+++ src/main/scala/ammonite/kernel/Parsers.scala
@@ -15,7 +15,7 @@ object Parsers {
   // For some reason Scala doesn't import this by default
   private val `_` = scalaparse.Scala.`_`
 
-  val ImportSplitter: P[Seq[ammonite.util.ImportTree]] = {
+  val ImportSplitter: P[Seq[ImportTree]] = {
     val IdParser = P((Id | `_`).!).map(
       s => if (s(0) == '`') s.drop(1).dropRight(1) else s
     )
@@ -26,13 +26,12 @@ object Parsers {
     )
     val Prefix = P(IdParser.rep(1, sep = "."))
     val Suffix = P("." ~/ (BulkImport | Selectors))
-    val ImportExpr: P[ammonite.util.ImportTree] = {
+    val ImportExpr: P[ImportTree] = {
       // Manually use `WL0` parser here, instead of relying on WhitespaceApi, as
       // we do not want the whitespace to be consumed even if the WL0 parser parses
       // to the end of the input (which is the default behavior for WhitespaceApi)
       P(Index ~~ Prefix ~~ (WL0 ~~ Suffix).? ~~ Index).map {
-        case (start, idSeq, selectors, end) =>
-          ammonite.util.ImportTree(idSeq, selectors, start, end)
+        case (start, idSeq, selectors, end) => ImportTree(idSeq, selectors, start, end)
       }
     }
     P(`import` ~/ ImportExpr.rep(1, sep = ",".~/))
